+++ src/models/studentModel.js
@@ -1,9 +1,10 @@
 const mongoose = require('mongoose');
 
 const studentData = new mongoose.Schema({
-  name:String,
+  firstName:String,
+  lastName:String,
   email:String,
-  password:String,
+  phoneNo:String,
 
 });