+++ src/routes/user.js
@@ -2,7 +2,11 @@ const express = require("express");
 const router = express.Router();
 require('dotenv').config();
 const studentService = require('../services/userService');
-const logger = require("../loggers/loggers");
+
+const { logger } = require("../loggers/loggers");
+//***************************************/
+//Routes - The API routes maps to the Controllers
+//***************************************/
 
 
 // Define the route for creating a student
@@ -32,7 +36,7 @@ router.post('/studentCreate', async (req, res) => {
 // Define the route for find student
 router.get('/studentGet',async(req, res) => {
 try {
-  const studentGetResult = await studentService.GetUser(req);
+  const studentGetResult = await studentService.getUser(req);
 
   // Respond with the result
   res.json(studentGetResult);
@@ -49,6 +53,50 @@ try {
     message: 'Something went wrong, please try again later',
   });
 }
+});
+
+// Define the route for update Data student
+router.put('/studentUpdate',async(req, res) => {
+  try {
+    const studentGetResult = await studentService.updateUser(req);
+  
+    // Respond with the result
+    res.json(studentGetResult);
+  } catch (error) {
+    logger.info(
+      `(Router:) /studentUpdate ,(File:) user.js, (Error:1)` +
+        " " +
+        `${error} ` +
+        " " +
+        new Date()
+    );
+    res.status(500).json({
+      status: 'error',
+      message: 'Something went wrong, please try again later',
+    });
+  }
+});
+
+// Define the route for Delete Data student
+router.delete('/studentDelete',async(req, res) => {
+  try {
+    const studentGetResult = await studentService.deleteUser(req);
+  
+    // Respond with the result
+    res.json(studentGetResult);
+  } catch (error) {
+    logger.info(
+      `(Router:) /studentDelete ,(File:) user.js, (Error:1)` +
+        " " +
+        `${error} ` +
+        " " +
+        new Date()
+    );
+    res.status(500).json({
+      status: 'error',
+      message: 'Something went wrong, please try again later',
+    });
+  }
 });
 
 module.exports = router;